backend/requirements.txt
Django==4.2
djangorestframework==3.14
mysqlclient==2.1.1
django-cors-headers==4.2.0
gunicorn==20.1.0
django-environ

elasticsearch-dsl
django-elasticsearch-dsl